requirements.txt
python-dotenv
numpy
matplotlib
pymongo
certifi
pymongo[srv]==3.12
scikit-learn
# -e .
